+++ src/locale/pl.ts
@@ -8,6 +8,7 @@ const translations = {
     settings: "Ustawienia",
     info: "Informacje",
     network: "Sieć",
+    access: "Dostęp",
     nodes: "Węzły",
     console: "Konsola",
     usb: "USB",